+++ .env
@@ -1,0 +1,2 @@
+NEXT_PUBLIC_SUPABASE_URL=https://feiahswsfrswwmifoiga.supabase.co
+NEXT_PUBLIC_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6ImZlaWFoc3dzZnJzd3dtaWZvaWdhIiwicm9sZSI6ImFub24iLCJpYXQiOjE2ODM2NjA2NjMsImV4cCI6MTk5OTIzNjY2M30.4X4enaKENmENR2UgJGGv-VpcBEYU_tfDqOuxOU2P4-Q
