+++ src/styles/Login.module.css
@@ -1,80 +1,123 @@
 .mainContainer {
-    background-color: #EBE5DF;
+    background-color: white;
     width: 100vw;
     max-width: 144rem;
     height: 100vh;
-    position: relative;
-    display: grid;
-    grid-template-columns: repeat(12,1fr);
-    grid-gap: 1rem;
-    grid-template-rows: repeat(12,1fr);
+    display: flex;
 }
 
-.logoTexto {
-    font-size: 5rem;
-    font-weight: bold;
-    position: relative;
-    color: #333333;
-    width: 18rem;
-    height: 11rem;
-    /* position: absolute; */
-    display: grid;
-    grid-column: 2;
-    grid-row: 2;
+.leftContainer{
+    display: flex;
+    flex-direction: column;
+    height: 100vh;
+    width: 100vw;
+    flex:1;
+    background-color: #b50909;
+    border-radius: 20px;
+    margin: 10px;
+    margin-top: 5px;
+    align-items: center;
 }
 
-.logoMais {
-    font-size: 10rem;
-    font-weight: bold;
-    position: absolute;
-    color: #9747FF;
-    top: 0rem;
-    left: 13rem;
+.logogradeplus{
+    fill: white;
+    width: 520px;
+    height: 380px;
+}
+.h1{
+    color: white;
+    text-align: center;
+    justify-content: center;
+    font-size: 60px;
+    margin-top: 0px;
+}  
+.p1{
+    color: white;
+    text-align: center;
+    justify-content: center;
+    font-size: 30px; 
+}
+
+.rightContainer{
+    /* border: solid red; */
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    gap: 2rem;
+    flex:1;
+    height: 100vh;
+}
+
+.Entrarbutton{
+    color: white;
+    border-radius: 15px;
+    width: 280px;
+    height: 65px;
+    margin: 10px auto;
+    background-color: #b50909;
+    border-style: solid;
+    border-color: white;
+    text-align: center;
+    justify-content: center;
+    font-size: 3rem;
+    line-height: 1px;
 }
 
 .textbox {
-    grid-column: 7/span8;
-    grid-row: 3;
+    width: 100%;
 }
 
 .textbox h1 {
-    font-size: 10rem;
-    color: #9747FF;
+    font-size: 8rem;
+    color: #b50909;
     line-height: 12rem;
-    font-weight: normal;
+    text-align: center;
 }
 
 .textbox p {
     font-size: 3rem;
     color: #707070;
-    padding: 0 1rem;
+   
+    text-align: center;
 }
 
 .formLogin {
-    /* border: solid; */
-    grid-column: 7/12;
-    grid-row: 5;
-    display: grid;
-    gap: 1rem;
-    /* padding: 2rem 0rem; */
+   display: flex;
+   flex-direction: column;
+   align-items: center;
+   gap: 1rem;
 }
 
 .formLogin input {
-    width: 47rem;
-    height: 6rem;
     background-color: #FFF;
     box-shadow: 1px 1px 0px 1px #333;
-    text-align: right;
+    text-align: left;
     box-sizing: border-box;
-    padding: 0 5rem;
+    padding: 0 0 0 8rem;
+}
+
+.wideInput{
+    width: 47rem;
+    height: 6rem;
+}
+.shortInput{
+    width: 23rem;
+    height: 6rem;
+}
+
+.setPassContainer{
+    display: flex;
+    gap: 1rem;
 }
 
 .formLogin input:hover {
     box-shadow: 2px 2px 0px 1px #333;
+    transition: box-shadow 0.5s ease;
 }
 
 .btnSubmitLogin {
-    background-color: #9747FF;
+    background-color: #b50909;
     color: #F0F0F0;
     font-size: 3rem;
     display: flex;
@@ -86,10 +129,11 @@
     box-shadow: 1px 1px 0px 1px #333;
 }
 
-.btnSubmitLogin:focus,
+/* .btnSubmitLogin:focus, */
 .btnSubmitLogin:hover {
-    background-color: #BA85FF;
+    background-color: #b50909;
     box-shadow: 2px 2px 0px 1px #333;
+    transition: box-shadow 0.5s ease;
 }
 
 .inputWrapper {
@@ -99,8 +143,8 @@
 
 .inputWrapper label {
     position: absolute;
-    bottom:0;
-    right: 0;
+    bottom: 0;
+    left: 0;
     padding: 1.5rem;
     font-size: 2rem;
     color: #333;
@@ -109,13 +153,55 @@
     transition: .5s;
 }
 
-.inputWrapper input:focus ~ label, 
-.inputWrapper input:valid ~ label {
+.inputWrapper input:focus ~ label ,
+.inputWrapper input:valid ~ label ,
+.inputWrapper input:not(:placeholder-shown) ~ label{
   top: 0;
-  right: 0;
+  left: 0;
   color: #333;
   font-size: 1.4rem;
-  padding: 0.5rem;
+  padding: 1rem;
+}
+
+.invalidInput{
+    border: solid #ff0000 1px;
+}
+
+.validInput{
+    border: solid #04AA6D 1px;
+}
+
+.errorMessage{
+    display: none;
+    color: #FF9999;
+    font-weight: lighter;
+    /* display: block; */
+    position: absolute;
+    top: 0;
+    right: 0;
+    margin: 1rem;
+}
+
+.emptyMessage{
+    display: none;
+    color: #FF9999;
+    font-weight: lighter;
+    /* display: block; */
+    position: absolute;
+    top: 0;
+    right: 0;
+    margin: 1rem;
+}
+
+.usedMessage{
+    display: none;
+    color: #FF9999;
+    font-weight: lighter;
+    /* display: block; */
+    position: absolute;
+    top: 0;
+    right: 0;
+    margin: 1rem;
 }
 
 .goSignUp{
@@ -147,4 +233,43 @@
 }
 .btnGoogle:hover{
     box-shadow: 2px 2px 0px 1px #333;
+}
+
+.setCheckPassWrapper{
+    display: flex;
+    gap: 1rem;
+}
+
+.passModal{
+    width: 15rem;
+    height: 20rem;
+    background-color: #333333;
+    position: absolute;
+    top: 0;
+    left: -16rem;
+    display: flex;
+    flex-direction: column;
+    justify-content: space-around;
+    align-items: center;
+    padding: 1rem;
+    text-align: center;
+    color: #707070;
+}
+
+.passModal span:first-child{
+    color: #F0F0F0;
+}
+
+.greenPass{
+    color: #04AA6D;
+}
+
+.containerEmailSent{
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    gap: 3rem;
+    font-size: 2rem;
+    color: #333;
 }
